requirements.txt
textarena>=0.1.0
anthropic>=0.18.0
python-dotenv>=1.0.0
requests>=2.31.0
nltk>=3.8.1
flask>=2.3.3
pandas>=2.0.0
notebook>=7.0.0
